editor n3names - adding

+++ anreader_sources/device/anames.h
@@ -25,7 +25,7 @@ public:
     virtual void setHidden(uint index, bool value = true) = 0;
     virtual void setUsed(uint index, bool value) = 0;
     virtual void setName(uint index, const QString value) = 0;
-    virtual void setFilled(const uchar value) = 0;
+    virtual void setFilled(const uint value) = 0;
 
 
     QByteArray& data() const {return m_data; }
